requirements.txt
asgiref
azure-common==1.1.28
azure-core==1.24.1
azure-storage-blob==12.0.0
azure-storage-common==2.1.0
backports.zoneinfo==0.2.1
boto3
botocore
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.4
Cython==0.29.30
Django
django-dotenv==1.4.2
django-environ==0.9.0
environ==1.0
idna==3.3
isodate==0.6.1
jmespath
msrest==0.7.1
oauthlib==3.2.0
parse==1.6.5
pycparser==2.21
pyparser==1.0
pyparsing==2.1.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests
requests-oauthlib==1.3.1
s3transfer
six==1.16.0
sqlparse==0.4.2
typing_extensions
urllib3==1.26.9
